CELE-63 Move some interfaces to the internal model
