+++ applications/visualizer/frontend/src/models/index.ts
@@ -1,0 +1,3 @@
+export { Workspace } from './workspace.ts';
+export type { NeuronGroup, Neuron, Dataset } from './models.ts';
+export { ViewMode } from './models.ts';
